make buttons work and look nice
renamed the file containing the input field and button, fixed the error upon passing the input to api, and made button appear on site
