+++ adversarial_apps/src/components/UpdateCompany.tsx
@@ -1,0 +1,103 @@
+"use client";
+
+import React, { useState } from "react";
+import { useRouter } from "next/navigation";
+import { toast } from "react-toastify";
+
+interface VerifyButtonProps
+{
+    cik: string;
+}
+
+export default function VerifyUpdate({ cik }: VerifyButtonProps)
+{
+    const router = useRouter();
+    const [loading, setLoading] = useState(false);
+    const [score, setScore] = useState<number | "">(""); // Ensure controlled input
+
+    const handleUpdate = async (e: React.FormEvent) =>
+    {
+        e.preventDefault(); // Prevent default form submission behavior
+
+        if (score === "" || isNaN(Number(score)))
+        {
+            toast.error("Please enter a valid risk score between 0 and 100.");
+            return;
+        }
+
+        setLoading(true);
+
+        try
+        {
+            const data = { action: "update_company_score", cik, risk_score: Number(score) }; // Use state value
+
+            console.log("Submitting Data:", data);
+
+            const response = await fetch(`/api/call-python-api`, {
+                method: "POST",
+                headers: {
+                    "Content-Type": "application/json",
+                },
+                body: JSON.stringify(data),
+            });
+
+            if (!response.ok)
+            {
+                throw new Error("Failed to update company score");
+            }
+
+            const result = await response.json();
+
+            if (result.status !== "success")
+            {
+                toast.error(`Error updating company: ${result.message || result.error}`);
+                return;
+            }
+
+            toast.success(result.message || "Company score updated successfully!");
+            router.refresh();
+        } catch (error)
+        {
+            console.error("Verification error:", error);
+            toast.error("Error updating company. Please try again later.");
+        } finally
+        {
+            setLoading(false);
+        }
+    };
+
+    return (
+        <form className="flex flex-col sm:flex-row items-end py-2 gap-4" onSubmit={handleUpdate}>
+            {/* Input Field Container */}
+            <div className="flex flex-col">
+                <label htmlFor="riskscore" className="block text-sm font-medium text-white">
+                    Risk Score:
+                </label>
+                <input
+                    type="number"
+                    id="riskscore"
+                    name="riskscore"
+                    min="0"
+                    max="100"
+                    value={score}
+                    onChange={(e) => setScore(e.target.value ? Number(e.target.value) : "")}
+                    className="w-64 px-4 py-2 mt-1 border rounded-md focus:ring-blue-500 focus:border-blue-500 text-gray-800"
+                    placeholder="Input score..."
+                    required
+                />
+            </div>
+
+            {/* Button Aligned to Bottom */}
+            <button
+                type="submit"
+                disabled={loading}
+                className={`px-6 py-2 bg-blue-900 text-white rounded-md shadow-md hover:bg-blue-600 transition-colors self-end ${loading ? "opacity-50 cursor-not-allowed" : ""
+                    }`}
+            >
+                {loading ? "Updating..." : "Update Company"}
+            </button>
+        </form>
+
+
+    );
+}
